refactored a bunch of file classes to be named consistently / added w_file_disk to quickly chunk data to and from the disk

+++ src/game_engine/engine.cpp
@@ -116,7 +116,7 @@ bool w_engine::init_game_engine( std::string_view game_name, int argc, char* arg
 			glfwSetWindowSize( engine->window->window, static_cast<int>( rc.w ), static_cast<int>( rc.h ) );
 			glfwSetWindowAspectRatio( engine->window->window,
 									  100,
-									  ( v_window_h / v_window_w ) * 100 );
+									  static_cast<int>( ( v_window_h / v_window_w ) * 100 ) );
 		}
 
 		// game
